implement checking lastRemovalDate on removal

+++ src/routes/course/manage.ts
@@ -80,11 +80,26 @@ function removePromise(courseID: string, studentID: string) {
   )
 }
 
+const getLastRemovalDateOfLatestSemester = async () => {
+  const queryResult = await db.semester.query(`
+    SELECT lastSubjectRemovalDate FROM semester
+    WHERE id = (SELECT id FROM semester ORDER BY semester.yearYear DESC, semester.semesterNumber DESC LIMIT 1)
+  `)
+  return queryResult[0].lastSubjectRemovalDate
+}
+
 router.post('/', async (req, res) => {
   console.log(req.body)
-  const e = await handleRemove(req.body)
+  if(req.query.forceRemove) {
+    const e = await handleRemove(req.body)
+  } else {
+    const lastRemovalDate = await getLastRemovalDateOfLatestSemester()
+    console.log('lastremovaldate', lastRemovalDate)
+    const now = new Date()
+    console.log(lastRemovalDate.getTime() > now.getTime() ? 'more than' : 'less than')
+  }
   // if (e.length > 0) res.render('')
-  res.redirect('/')
+  res.redirect(req.baseUrl)
 })
 
 export default router
